requirements.txt
requests
pandas
openpyxl
office365-rest-python-client